+++ actionmailer/CHANGELOG.md
@@ -1,3 +1,10 @@
+*   Raise an exception when attachments are added after `mail` was called.
+    This is a safeguard to prevent invalid emails.
+
+    Fixes #16163.
+
+    *Yves Senn*
+
 *   Add `config.action_mailer.show_previews` configuration option.
 
     This config option can be used to enable the mail preview in environments